Removed .done()
Clint said that this .done() is an artifact from the old q promised (per https://dropship.atlassian.net/browse/LEO-171) - he said "the old version of promises we used was "q" and it required a .done() to terminate the chain. The build in Promises in node don't require the .done()"

Therefore, I don't think we need it, and it is breaking us.